## Changes Integrated:
- "New Discovery" bug fix (symbol='NEW' now updates)
- DOUBLE PRECISION type fixes in database.js
- K-Score interval changed to 12 hours (RPC savings)
- tokenQueue.js for queue-based token processing
- priceProvider.js and rpcMonitor.js services
- Holder count caching (5min TTL)
- Rate limiting for token indexing
- 8 troubleshooting documentation files

## Our Features Preserved:
- Harmony v2.0 with complete E-Score implementation
- Helius webhook for new token discovery
- API credit explosion fix (TTL + delta mode)

## Conflict Resolution:
- listener_worker.js: Combined policies (Helius webhook + CA search fallback)
- kScoreUpdater.js: Delta mode works in all modes, restricted to verified tokens

+++ src/migrations/003_distributed_polling.sql
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS polling_tasks (
     completed_at BIGINT,
     attempts INTEGER DEFAULT 0,
     last_error TEXT,
-    k_score_result INTEGER
+    k_score_result DOUBLE PRECISION
 );
 
 -- Partial unique index: only one pending task per mint
@@ -43,7 +43,7 @@ CREATE TABLE IF NOT EXISTS token_verifications (
     id SERIAL PRIMARY KEY,
     mint TEXT NOT NULL,
     node_id TEXT NOT NULL,
-    k_score INTEGER,
+    k_score DOUBLE PRECISION,
     node_signature TEXT,
     verified_at BIGINT NOT NULL,
 
@@ -83,7 +83,7 @@ CREATE INDEX IF NOT EXISTS idx_node_work_history_mint
 -- ============================================
 CREATE TABLE IF NOT EXISTS consensus_snapshots (
     mint TEXT PRIMARY KEY,
-    k_score_consensus INTEGER,
+    k_score_consensus DOUBLE PRECISION,
     agreeing_nodes INTEGER DEFAULT 0,
     total_nodes INTEGER DEFAULT 0,
     consensus_at BIGINT,
